(feat) Finish author tag query controller, and refactor for posts and authors to share the method
